==> 01_input-checks.Rmd <==
---
title: "Vstupní kontrola úplnosti a konzistence dat"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(metathis)
library(tidyverse)
library(arrow)
library(fontawesome)
library(pointblank)

source(here::here("shared.R"))
```

```{r load data}
dt <- read_parquet(here::here("data-processed", "misto_renamed-n.parquet"))
```

## Jak to číst

Tento skript/dokument funguje tak, že se snaží najít řádky/projekty s nějakou vnitřní nekonzistencí. V prvním gardu pak vyhodí počet takových řádků. Pokud je pod kódem vidět `0`, je vše v pořádku.

```{r}
al <- action_levels(warn_at = 1)
as <- action_levels(stop_at = 1)
```


## Počty a ID projektů

### Máme všude ID projektu?

```{r no-id, validate = F}
dt %>% 
  create_agent(label = "project IDs") %>% 
  col_vals_not_null(c(op_id, prj_id)) %>% 
  interrogate(extract_failed = F)
```

### Dávají čísla projektů smysl?

Délky:


```{r}
op_ids <- c("OP Z", "OP D", "OP VVV", "OP PIK", 
            "OP TP", "OP PPR", "OP ZP", "IROP")

op_ids %in% dt$op_id
```


```{r, validate = F}
dt %>% 
  create_agent(label = "project number", actions = al) %>% 
  col_is_character(starts_with("g_")) %>%
  col_vals_in_set(op_id, op_ids) %>%
  col_vals_regex(prj_id, ".{12,15}", brief = "Proj num len") %>% 
  col_vals_not_null(c(op_id, prj_id)) %>% 
  col_vals_regex(g_zuj_id, ".{11}", na_pass = T) %>% 
  interrogate(extract_failed = F)
```


### Počty řádků a projektů

Kolik je unikátních projektů

```{r distinct projects}
dt %>%
  summarise(prj_count = n_distinct(prj_id))
```

```{r radky-projekty}
dt %>% 
  group_by(op_id) %>% 
  summarise(n_radky = n(), 
            n_projekty = length(unique(prj_id)))
```

## Údaje o místě realizace

```{r dt-long-geo}

dt_long_g <- make_long_geo(dt)

write_parquet(dt_long_g %>% 
                drop_na(value), here::here("data-processed", "misto_long-geo.parquet"))
```

### Existují projekty bez jakékoli idenfitikace místa?

```{r noplace-n}
dt_noplace <- dt_long_g %>%
  group_by(prj_id, op_id, p_nazev, p_sidlo_nazev, prj_nazev) %>% 
  summarise(all_na = all(is.na(value))) %>% 
  ungroup() 
```

```{r noplace-agent, message=F}
dtnopl_agent <- dt_noplace %>% 
  create_agent() %>% 
  col_vals_equal(all_na, FALSE)
```


```{r noplace-validate, message=F, validate = F}
dtnopl_agent %>% 
  interrogate(extract_failed = F)
```


```{r noplace-faillist, message=F}
dtnopl_agent %>% 
  interrogate() %>% 
  get_data_extracts(i = 1)
```


```{r noplace-op}
dt_noplace %>% 
  filter(all_na) %>% 
  count(op_id)
```


### Chybí někde ID tam, kde je název a naopak?


```{r missing-corresp}
dt_misscorresp <- dt_long_g %>% 
  select(-prj_anotace, -starts_with("z_sidlo")) %>% 
  pivot_wider(id_cols = c(prj_id, prj_radek, level), 
              names_from = typ, values_from = value)
```

```{r missing-corresp-n}
dt_misscorresp %>% 
  filter(is.na(id) != is.na(nazev)) %>% 
  nrow()
```

### Jsou ID a názvy konzistentní?


```{r id-name-mismatch-n}
dt_long_g %>% 
  select(-prj_anotace, -starts_with("z_sidlo")) %>% 
  pivot_wider(id_cols = c(prj_id, prj_radek, level), 
              names_from = typ, values_from = value) %>% 
  group_by(level, id) %>%
  summarise(n_names_in_id = length(unique(nazev))) %>% 
  filter(n_names_in_id > 1) %>% 
  nrow()
```



## Dávají data o příjemci smysl?



### Je ve všech řádcích každého projektu to stejné sídlo? 

```{r prij-sidlo-id-mismatch, validate = F}
dt %>% 
  group_by(prj_id) %>% 
  summarise(sidlo_id_unique = length(unique(p_sidlo_id)),
            sidlo_name_unique = length(unique(p_sidlo_nazev))) %>% 
  create_agent() %>% 
  col_vals_equal(sidlo_id_unique, 1) %>% 
  col_vals_equal(sidlo_name_unique, 1) %>% 
  interrogate(extract_failed = F)
```


### Chybí někde údaje o příjemci?

```{r prij-missing}
dt %>% 
  filter(prj_radek == 1) %>% 
  filter(is.na(p_nazev) | is.na(p_ico)) %>% 
  select(op_id, prj_id, p_sidlo_id, p_sidlo_nazev, p_nazev, p_ico, prj_nazev)
```

### Chybí někde údaje o sídle?

```{r prij-sidlo-missing}
dt %>% 
  filter(prj_radek == 1) %>% 
  filter(is.na(p_sidlo_id) | is.na(p_sidlo_nazev)) %>% 
  select(op_id, prj_id, p_sidlo_id, p_sidlo_nazev, p_nazev, prj_nazev)
```

Ehm, takže asi existují projekty, kde známe místo sídla a neznáme název příjemce

```{r prij-incognito}
dt %>% 
  filter(prj_radek == 1) %>% 
  filter(is.na(p_nazev) & (!is.na(p_sidlo_id) | !is.na(p_sidlo_nazev))) %>% 
  select(op_id, p_sidlo_id, p_sidlo_nazev, p_nazev, p_ico, prj_id, prj_nazev) %>% 
  arrange(prj_id)
```

### Jsou údaje o sídle vnitřně konzistentní?

```{r prij-sidlo-match-df}
dt_long_sidla <- dt %>%
  group_by(prj_id, prj_radek, p_nazev, p_ico) %>% 
  pivot_longer(cols = starts_with("p_sidlo_")) %>%
  separate(name, c("p_", "sidlo", "typ")) %>% 
  select(op_id, prj_id, typ, value)
```

```{r prij-sidlo-mismatch-list, validate = F}
dt_long_sidla %>% 
  pivot_wider(id_cols = c(prj_id, prj_radek),
              names_from = typ, values_from = value) %>% 
  group_by(prj_id) %>% 
  summarise(distinct_ids = length(unique(id)),
            distinct_names = length(unique(nazev))) %>% 
  create_agent() %>% 
  col_vals_equal(distinct_ids, 1, 
                 brief = "Each prj only has prijemce in one place ID") %>% 
  col_vals_equal(distinct_names, 1, 
                 brief = "Each prj only has prijemce in one place name") %>% 
  interrogate(extract_failed = F)
```

Do recipients move seat? Change names?

```{r prij-sidlo-change, validate = F}
dt %>% 
  group_by(p_ico) %>% 
  summarise(distinct_seats = length(unique(p_sidlo_id)),
            distinct_names = length(unique(p_nazev))) %>%  
  create_agent() %>% 
  col_vals_equal(distinct_seats, 1, brief = "Each recipient only has one seat") %>% 
  col_vals_equal(distinct_names, 1, brief = "Each recipient only has one name") %>% 
  interrogate(extract_failed = F)
```

## Údaje o projektech v CHKO/NP

```{r chko-load}
prj_chkonp <- read_parquet(here::here("data-processed", "prj_chkonp.parquet"))
```

```{r chko-validate}
prj_chkonp %>% 
  create_agent(actions = action_levels(warn_at = 1, stop_at = .3)) %>% 
  col_vals_in_set(prj_id, unique(dt$prj_id)) %>% 
  interrogate()
```
