==> 02_validate.Rmd <==
---
title: "Kontrola proti číselníkům"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(arrow)
library(pointblank)

source(here::here("shared.R"))
```

```{r data}
dt <- read_parquet(here::here("data-processed", "misto_renamed-n.parquet"))
dtl <- read_parquet(here::here("data-processed", "misto_long-geo.parquet"))
```

```{r}
setdiff(dt$prj_id, dtl$prj_id)
```


```{r codes-in-data}
obec_id_e <- unique(dt$g_obec_id)
kraj_id_e <- unique(dt$g_kraj_id)
orp_id_e <- unique(dt$g_orp_id)
okres_id_e <- unique(dt$g_okres_id)
zuj_id_e <- unique(dt$g_zuj_id)
```


### Jak vypadají kódy v datech?

```{r sample codes}
head(obec_id_e)
head(zuj_id_e)
head(orp_id_e)
head(okres_id_e)
head(kraj_id_e)
```

### Podle toho načteme a upravíme metadata

Děje se v `03_load-metadata.R`, aby se stahování a počítání dělalo jen jednou.

```{r metadata-load}
source("read_metadata.R")
```

```{r metadata-prep}
obce_n <- obecnuts %>% 
  mutate(kod_n = paste0(CZNUTS, CHODNOTA2))

zuj_n <- zujnuts %>% 
  mutate(kod_n = paste0(CZNUTS, CHODNOTA2))
  
head(obce_n$kod_n)
head(zuj_n$kod_n)
```

## Existují kódy uvedené v datech v oficiálních číselnících?

```{r validated}
dt_validated <- dt %>% 
  mutate(valid_kraj = g_kraj_id %in% csu_kraj$CZNUTS,
         valid_obec = g_obec_id %in% obce_n$kod_n,
         valid_orp = g_orp_id %in% csu_orp$CHODNOTA,
         valid_okres = g_okres_id %in% csu_okres$CHODNOTA,
         valid_zuj = g_zuj_id %in% zuj_n$kod_n) %>% 
  mutate(valid_kraj = if_else(is.na(g_kraj_id), NA, valid_kraj),
         valid_obec = if_else(is.na(g_obec_id), NA, valid_obec),
         valid_orp = if_else(is.na(g_orp_id), NA, valid_orp),
         valid_okres = if_else(is.na(g_okres_id), NA, valid_okres),
         valid_zuj = if_else(is.na(g_zuj_id), NA, valid_zuj)) %>% 
  select(matches(c("valid", "g_.*_id")))

skimr::skim(dt_validated)
```

Něco se děje s kódy ZÚJ: část nesedí s číselníkem - a část je taky kratší.

```{r zuj-len}
dt %>% 
  mutate(zuj_len = str_length(g_zuj_id)) %>% 
  count(op_id, zuj_len) %>% 
  spread(zuj_len, n)
```

```{r zuj-short}
dt %>% 
  filter(str_length(g_zuj_id) < 11) %>% 
  distinct(op_id, g_zuj_id) 
```
Což jsou podle všeho městské části Prahy (je jich 57).

Oprava kódů ZÚJ: ke kratším na začátek přilepíme kód kraje

```{r zuj-fix}
dt_zujfix <- dt %>% 
  mutate(g_zuj_shortid = str_sub(g_zuj_id, -6, -1)) %>% 
  left_join(zujnuts %>% 
              select(g_zuj_shortid = CHODNOTA2, CZNUTS)) %>% 
  mutate(g_zuj_id_new = if_else(str_length(g_zuj_id) < 11,
                                paste0(CZNUTS, g_zuj_shortid), g_zuj_id),
         g_zuj_id = g_zuj_id_new,
         g_obec_shortid = str_sub(g_obec_id, -6, -1)) %>% 
  select(-g_zuj_id_new, -CZNUTS)

dt_zujfix %>% 
  mutate(ln = str_length(g_zuj_id)) %>% 
  count(ln)
```

Kontrola znova:

```{r revalidate}
dt_zujfix_validated <- dt_zujfix %>% 
  mutate(valid_kraj = g_kraj_id %in% csu_kraj$CZNUTS,
         valid_obec = g_obec_id %in% obce_n$kod_n,
         valid_orp = g_orp_id %in% csu_orp$CHODNOTA,
         valid_okres = g_okres_id %in% csu_okres$CHODNOTA,
         valid_zuj = g_zuj_id %in% zuj_n$kod_n) %>% 
  mutate(valid_kraj = if_else(is.na(g_kraj_id), NA, valid_kraj),
         valid_obec = if_else(is.na(g_obec_id), NA, valid_obec),
         valid_orp = if_else(is.na(g_orp_id), NA, valid_orp),
         valid_okres = if_else(is.na(g_okres_id), NA, valid_okres),
         valid_zuj = if_else(is.na(g_zuj_id), NA, valid_zuj)) %>% 
  select(matches(c("valid", "g_.*_id")))

skimr::skim(dt_zujfix_validated)
```

Fajn: teď už to sedí: podíly chybějících jsou stejné v původních proměnních jako ve validačních, plus máme velmi málo chybných kódů - jen 2 kraje nesedí.

Které?

```{r polsko}
kraj_id_e[!kraj_id_e %in% csu_kraj$CZNUTS]
```

Polský region Bielski a `NA`.

```{r polsko-list}
dt %>% 
  filter(prj_id == dt %>% 
           filter(g_kraj_id == "PL225") %>% 
           pull(prj_id))
```

Hmm, jeden projekt OP Z, který se odehrával jen v Polsku... Ale je to výzva na mezinárodní mobilitu, takže asi OK - prostě se jen vyřadí.

```{r fixed-write}
write_parquet(dt_zujfix, here::here("data-processed", "misto_fix-01-zuj.parquet"))
```

```{r fixed-long-write}
dt_fixed_long <- make_long_geo(dt_zujfix)
write_parquet(dt_fixed_long %>% 
                drop_na(value), 
              here::here("data-processed", "misto_fix-01-zuj_long-geo.parquet"))
```

### Obce

```{r}
obce_nomatch <- obec_id_e[!obec_id_e %in% paste0(obecnuts$CZNUTS, obecnuts$CHODNOTA2)]
```

```{r}
obecnuts %>% 
  mutate(id = paste0(CZNUTS, CHODNOTA2)) %>% 
  filter(id %in% obce_nomatch[2]) %>% 
  pull(TEXT2)
```
 
 Brdy: https://www.risy.cz/cs/vyhledavace/obce/539996-brdy

## Validita jmen proti registrům

```{r geoid-validation-merge}
dt_val_ids <- dt_zujfix %>% 
  # select(prj_id, starts_with("g_")) %>% 
  left_join(zuj_n %>% 
              select(g_zuj_nazev_csu = TEXT2, 
                     g_zuj_id = kod_n)) %>% 
  left_join(obce_n %>% 
              select(g_obec_nazev_csu = TEXT2, 
                     g_obec_id = kod_n)) %>% 
  left_join(csu_okres %>% 
              select(g_okres_nazev_csu = TEXT, 
                     g_okres_id = CHODNOTA)) %>% 
  left_join(csu_orp %>% 
              select(g_orp_nazev_csu = TEXT, 
                     g_orp_id = CHODNOTA)) %>% 
  left_join(csu_kraj %>% 
              select(g_kraj_nazev_csu = TEXT, 
                     g_kraj_id = CZNUTS))
```

```{r geoid-validation-validate, validate = F}
dt_val_ids_agent <- dt_val_ids %>% 
  mutate(across(starts_with("g_"), tolower)) %>% 
  create_agent() %>% 
  col_vals_equal(g_zuj_nazev, 
                 vars(g_zuj_nazev_csu), na_pass = T) %>% 
  col_vals_equal(g_obec_nazev, 
                 vars(g_obec_nazev_csu), na_pass = T) %>% 
  col_vals_equal(g_orp_nazev, 
                 vars(g_orp_nazev_csu), na_pass = T) %>% 
  col_vals_equal(g_okres_nazev, 
                 vars(g_okres_nazev_csu), na_pass = T) %>% 
  col_vals_equal(g_kraj_nazev, 
                 vars(g_kraj_nazev_csu), na_pass = T)
```

```{r geoid-validation-interrogate, validate = F}
dt_val_ids_agent %>% interrogate(extract_failed = F)
```

```{r geoid-validation-inspect-1}
dt_val_ids_agent %>% 
  interrogate() %>% 
  get_data_extracts(1) %>% 
  select(g_zuj_id, g_zuj_nazev, g_zuj_nazev_csu)
```
```{r geoid-validation-inspect-2}
dt_val_ids_agent %>% 
  interrogate() %>% 
  get_data_extracts(2) %>% 
  select(g_obec_id, g_obec_nazev, g_obec_nazev_csu) %>% 
  distinct()
```
```{r geoid-validation-rename-cols}
dt_val_ids_csunames <- dt_val_ids %>% 
  select(-g_zuj_nazev, -g_obec_nazev) %>% 
  rename(g_zuj_nazev = g_zuj_nazev_csu,
         g_obec_nazev = g_obec_nazev_csu) %>% 
  select_at(vars(-ends_with("_csu")))

dt_val_ids_csunames_long <- make_long_geo(dt_val_ids_csunames) %>% 
  drop_na(value)
```

```{r geoid-validation-write}
write_parquet(dt_val_ids_csunames, 
              here::here("data-processed", 
                         "misto_fix-02-gnames.parquet"))
write_parquet(dt_val_ids_csunames_long, 
              here::here("data-processed",
                         "misto_fix-02-gnames_long-geo.parquet"))
```
