==> 10_compile-export.Rmd <==
---
title: "Kompilace dat"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(arrow)

cnf <- config::get()
```

```{r load parts}
schema <- schema(prj_id = string(),
                 radek = int32(),
                 op_id = string(),
                 level = dictionary(ordered = T),
                 geo_id = string(),
                 level_orig = string(),
                 geo_id_orig = string(),
                 rozpad_duvod = string(),
                 rozpad_typ = string(),
                 obec_puvod = string(),
                 chunk = int32(),
                 geo_id_long = string())

ds_stubs <- c("dtl_resolved_placeless", "dtl_resolved_smpl",
              "dtl-obecless-rozpadnute")

file_nh_resolved_previously <- here::here("data-processed", "dtl-nehier", "all.parquet")
if(file.exists(file_nh_resolved_previously)) {
  ds_stubs <- c(ds_stubs, "dtl-nehier")
}

dt_list <- purrr::map(ds_stubs, 
                      ~open_dataset(here::here("data-processed", .x), 
                                    schema = schema))
```

```{r load all into one}
dt_all <- open_dataset(dt_list)
dim(dt_all)
```

```{r cleanup previous export}
try(fs::dir_delete(here::here(cnf$arrow_output_dir)), silent = T)
```

```{r resave}
write_dataset(dt_all, path = here::here(cnf$arrow_output_dir), 
              partitioning = cnf$arrow_partitioning, 
              format = "parquet")
```

```{r dogfood, eval=T}
dss <- open_dataset(cnf$arrow_output_dir)
```

```{r check-dims}
dim(dss)
```

```{r check-schema}
dss$schema
```

```{r check head}
dss %>% head() %>% collect()
```


```{r check chunks, eval=F}
dss %>% 
  select(chunk, op_id) %>% 
  collect() %>% 
  count(chunk, op_id)
```

```{r check grouping, eval=FALSE}
dss %>% 
  select(obec_puvod, rozpad_typ, op_id) %>% 
  collect() %>% 
  distinct(obec_puvod, rozpad_typ, op_id)
```


```{r sample projects}
prj_smpl <- dss %>% 
  select(prj_id, op_id, rozpad_typ, obec_puvod) %>% 
  collect() %>% 
  distinct() %>% 
  group_by(op_id, rozpad_typ, obec_puvod) %>% 
  sample_n(min(n(), 10))
```

```{r sample rows}
dss_sample <- dss %>% 
  filter(prj_id %in% prj_smpl$prj_id) %>% 
  collect()
```

```{r write sample general}
writexl::write_xlsx(dss_sample, here::here("sample_export.xlsx"))
```


```{r write sample opz}
writexl::write_xlsx(dss %>% 
                      filter(op_id == "OP Z" & chunk == 1) %>% 
                      collect(), here::here("opz-one-chunk.xlsx"))
```

```{r, eval=F}
dss_all <- dss %>% collect()
```
